big refator to ui element's own service
